src/thrift/mpes-llm/requirements.txt
fastapi
uvicorn
pydantic
httpx
llama-cpp-python
numpy
torch
thriftpy2